rag-news-ingestion/requirements.txt
feedparser
requests
beautifulsoup4
newspaper3k
chromadb
langchain-community
requests
google-generativeai
